+++ site/components/module/articleHeader.tsx
@@ -7,7 +7,7 @@ export type ArticleHeaderArgument = {
 };
 
 export function articleHeader(store: Store): HComponentFn<ArticleHeaderArgument> {
-    const ArticleHeader = element(store, "header", { name: "article-header" });
+    const ArticleHeader = element(store, { tag: "header", name: "article-header" });
     const Title = element(store);
     const Meta = element(store);
 
@@ -27,10 +27,10 @@ export function articleHeader(store: Store): HComponentFn<ArticleHeaderArgument>
 
     registerStyle(store, ArticleHeader, component_styles);
 
-    return component(ArticleHeader, ({ title }, ...child) => (
+    return component(ArticleHeader, ({ title }, ...children) => (
         <ArticleHeader>
             <Title>{title}</Title>
-            <Meta>{child}</Meta>
+            <Meta>{children}</Meta>
         </ArticleHeader>
     ));
 }