+++ src/lib/core/elements.ts
@@ -1943,9 +1943,14 @@ export function printDefine() {
 }
 
 export function gt<K extends Tag | qrillTag>(tag: K): HElementFn<K> {
-    return {
-        [tag]: (attribute: AttributeOf<K>, ...child: HNode[]) => ({ tag, attribute, child }),
-    }[tag];
+    return Object.assign(
+        {
+            [tag]: (attribute: AttributeOf<K>, ...children: HNode[]) => ({ tag, attribute, children }),
+        }[tag],
+        {
+            designator: tag,
+        },
+    );
 }
 
 export function DOCTYPE(): string {