+++ src/page/page.tsx
@@ -1,4 +1,4 @@
-import { element } from "@/lib/core/component";
+import { component, element } from "@/lib/core/component";
 import type { HArgument, HComponentFn } from "@/lib/core/component";
 import type { Store } from "@/lib/core/store";
 
@@ -9,11 +9,11 @@ const site = {
 };
 
 export function page(store: Store): HComponentFn<HArgument> {
-    const PageHeader = element(store, "header", { class: ["page-header", "content"] });
-    const PageFooter = element(store, "footer", { class: "page-footer" });
-    const PageFooterCopyright = element(store, "div", { class: "page-footer-copyright" });
+    const PageHeader = element(store, { tag: "header", class: ["page-header", "content"] });
+    const PageFooter = element(store, { tag: "footer", class: "page-footer" });
+    const PageFooterCopyright = element(store, { class: "page-footer-copyright" });
 
-    return (_attribute, ...child) => (
+    return component("html", (_attribute, ...children) => (
         <html lang={site.lang}>
             <head>
                 <meta charset="utf-8" />
@@ -28,11 +28,11 @@ export function page(store: Store): HComponentFn<HArgument> {
                         <a href="/">{site.name}</a>
                     </h1>
                 </PageHeader>
-                {child}
+                {children}
                 <PageFooter>
                     <PageFooterCopyright>&copy; 2025 {site.name}</PageFooterCopyright>
                 </PageFooter>
             </body>
         </html>
-    );
+    ));
 }